% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{compute_beta}
\alias{compute_beta}
\title{Function which computes the coefficients (denoted beta) for the recursive formulas
in determining K from Bodnar et. al. 2021}
\usage{
compute_beta(i, j, Psi)
}
\arguments{
\item{i}{integer a integer greater than one.}

\item{j}{integer a integer greater than one.}

\item{Psi}{vector of shrinkage coefficients.}
}
\value{

}
\description{
Function which computes the coefficients (denoted beta) for the recursive formulas
in determining K from Bodnar et. al. 2021
}
