% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{ComputeBeta}
\alias{ComputeBeta}
\title{A helper function for computation of coefficients.}
\usage{
ComputeBeta(i, j, Psi)
}
\arguments{
\item{i}{integer a integer greater than one.}

\item{j}{integer a integer greater than one.}

\item{Psi}{vector of shrinkage coefficients.}
}
\value{
a number
}
\description{
Function which computes the coefficients (denoted beta) for the recursive formulas in determining K from Bodnar et. al. 2021
}
