% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{ComputeD}
\alias{ComputeD}
\title{A small helper function which computes the terms D_{ij} from Bodnar et al. 2021}
\usage{
ComputeD(Ci, Cj)
}
\arguments{
\item{Ci}{numeric concentration ratio of period i}

\item{Cj}{numeric concentration ratio of period j}
}
\value{
a number
}
\description{
A small helper function which computes the terms D_{ij} from Bodnar et al. 2021
}
