% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConvexCombination}
\alias{ConvexCombination}
\title{Computes a convex combination between two vectors.}
\usage{
ConvexCombination(x1, x2, lambda)
}
\arguments{
\item{x1}{a vector.}

\item{x2}{a vector.}

\item{lambda}{a numeric value between 0 and 1.}
}
\value{
a vector
}
\description{
Computes a convex combination between two vectors.
}
\keyword{internal}
