% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulates_covariance.R
\name{generate_sigma}
\alias{generate_sigma}
\title{Simulates a covariance matrix}
\usage{
generate_sigma(p)
}
\arguments{
\item{p}{an integer the dimension on the covariance matrix}
}
\value{
a positive definite covariance matrix which has a specific eigenvalue distribution.
}
\description{
The simulated covariance matrix has eigenvalues at 0.2, 1 and 5 with a multiplicity of p/5, 2p/5 and 2p/5 respectively.
}
