% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSPortfolio.R
\name{new_DOSPortfolio}
\alias{new_DOSPortfolio}
\title{Constructor for the DOSPortfolio class.}
\usage{
new_DOSPortfolio(data, change_points, target_w, relative_loss, shrinkage_type)
}
\arguments{
\item{data}{a matrix of size (n x p), where n>p, containing, for instance, log-returns.}

\item{change_points}{a vector of break points. The breakpoints are what determines
when we recompute weights.}

\item{target_w}{a vector which is the target weights that one wants to shrink to in the first period.}

\item{relative_loss}{a numeric of the initial value of the relative loss for the variance of the GMV portfolio.}

\item{shrinkage_type}{the type of shrinkage estimator to use. The two implemented are "non-overlapping" and "overlapping"}
}
\value{
a DOSPortfolio class.
}
\description{
Constructor for the DOSPortfolio class.
}
