% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OnesVec}
\alias{OnesVec}
\title{Creates a vector of length p which has ones as elements.}
\usage{
OnesVec(p)
}
\arguments{
\item{p}{integer}
}
\value{
vector
}
\description{
Creates a vector of length p which has ones as elements.
}
