% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r0_strategy}
\alias{r0_strategy}
\title{Computes a simple plugin init for the computation of}
\usage{
r0_strategy(S_inv, S, b, c)
}
\arguments{
\item{lambda}{numeric}

\item{x1}{vector}

\item{x2}{vector}
}
\value{
vector
}
\description{
Computes a simple plugin init for the computation of
}
