% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r0Strategy}
\alias{r0Strategy}
\title{Simple plugin estimator for the relative loss. This function computes a
simple plugin estimator for the relative loss to the variance of the GMV portfolio
and the target portfolio.}
\usage{
r0Strategy(S_inv, S, target_portfolio, c)
}
\arguments{
\item{S_inv}{the inverse of a sample covariance matrix.}

\item{S}{a sample covariance matrix.}

\item{target_portfolio}{a vector representing a target portfolio which is shrunk to.}

\item{c}{a numeric which is the concentration ratio.}
}
\value{
vector
}
\description{
Simple plugin estimator for the relative loss. This function computes a
simple plugin estimator for the relative loss to the variance of the GMV portfolio
and the target portfolio.
}
