% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overlapping.R
\name{r_update}
\alias{r_update}
\title{Function which updates values of r0.}
\usage{
r_update(xi, c, r_prev)
}
\arguments{
\item{xi}{numeric in range of (0,1)}

\item{c}{numeric in range of (0,1)}

\item{r_prev}{numeric in range of (0,1)}
}
\value{
numeric
}
\description{
Function which updates values of r0.
}
\examples{
shrinkage_coef(1, 0.5)
}
