% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{rUpdateOverlapping}
\alias{rUpdateOverlapping}
\title{The recursive scheme for updating the relative loss in the overlapping scheme.}
\usage{
rUpdateOverlapping(Psi, c, prev_R, K)
}
\arguments{
\item{Psi}{vector of shrinkage coefficients.}

\item{c}{numeric between 0 and 1. It is the concentration ration.}

\item{prev_R}{numeric greater than 0, The previous value of the relative loss}

\item{K}{numeric parameter.}
}
\value{
a number
}
\description{
The recursive scheme for updating the relative loss in the overlapping scheme.
}
