% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonOverlapping.R
\name{shrinkage_coef}
\alias{shrinkage_coef}
\title{Computes the shrinkage for non-overlapping samples.}
\usage{
shrinkage_coef(r_prev, c)
}
\arguments{
\item{r_prev}{numeric}

\item{c}{numeric}
}
\value{
numeric
}
\description{
Computes the shrinkage for non-overlapping samples.
}
\examples{
shrinkage_coef(1, 0.5)
}
