% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overlapping.R
\name{shrinkage_coef_non_overlapping}
\alias{shrinkage_coef_non_overlapping}
\title{Computes the shrinkage coefficient for non-overlapping samples.}
\usage{
shrinkage_coef_non_overlapping(r_prev, c)
}
\arguments{
\item{r_prev}{numeric}

\item{c}{numeric}
}
\value{
numeric
}
\description{
Computes the shrinkage coefficient for non-overlapping samples.
}
