% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overlapping.R
\name{ShrinkageCoefNonOverlapping}
\alias{ShrinkageCoefNonOverlapping}
\title{Computes the shrinkage coefficient for non-overlapping samples.}
\usage{
ShrinkageCoefNonOverlapping(r_prev, c)
}
\arguments{
\item{r_prev}{numeric}

\item{c}{numeric}
}
\value{
numeric
}
\description{
Computes the shrinkage coefficient for non-overlapping samples.
}
