% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSPortfolio.R
\name{validate_input}
\alias{validate_input}
\title{Validates input to the DOSPortfolio function.}
\usage{
validate_input(data, change_points, target_w, relative_loss, shrinkage_type)
}
\arguments{
\item{data}{the data to validated, should be on long format.}

\item{change_points}{a vector of break points.}

\item{target_w}{the target vector at time point 0.}

\item{relative_loss}{the relative loss towards the GMV portfolios variance.}

\item{shrinkage_type}{what type of shrinkage method to be used.}
}
\value{
NULL, only called for its side effects
}
\description{
This function is exists to validate the assumptions made to derive the analytic formulas implemented in the different
functions. Is called for its side-effects.
}
