% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{w_gmv}
\alias{w_gmv}
\title{Computes the Global Minimum Variance (GMV) portfolio weights.}
\usage{
w_gmv(S_inv)
}
\arguments{
\item{r_prev}{numeric}

\item{c}{numeric}
}
\value{
vector
}
\description{
Computes the Global Minimum Variance (GMV) portfolio weights.
}
