% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overlapping.R
\name{w_non_overlapping}
\alias{w_non_overlapping}
\title{Computes the shrunk GMV portfolio weights}
\usage{
w_non_overlapping(data, w_random_target, r_previous)
}
\arguments{
\item{data}{data.frame on long format. Contains log returns.}

\item{r_previous}{numeric}

\item{w_target}{vector}
}
\value{
vector
}
\description{
Computes the shrunk GMV portfolio weights
}
