% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{w_overlapping}
\alias{w_overlapping}
\title{Computes the GMV weights when shrinking dynamically with overlapping samples.}
\usage{
w_overlapping(data, break_points, target_w, R)
}
\arguments{
\item{data}{a matrix in long format containing, for instance, log-returns.}

\item{break_points}{a vector of break points. The breakpoints are what determines
when we recompute weights.}

\item{target_w}{a vector which is the target weights that one wants to shrink to in the first period.}

\item{R}{a numeric of the initial value of the relative loss for the variance of the GMV portfolio.}
}
\value{
vector of portfolio weights
}
\description{
Computes the GMV weights when shrinking dynamically with overlapping samples.
}
\examples{
n <- 200*2
p <- 80
c <- p/n
break_points <- c(199)
data <- matrix(rt(n*p, df=5), ncol=p, nrow=n)
target_w <- as.vector(rep(1,p))/p
w_overlapping(data, break_points, target_w, 1)

}
