% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wGMV}
\alias{wGMV}
\title{Sample estimator of the weights of the global minimum variance portfolio}
\usage{
wGMV(data)
}
\arguments{
\item{S_inv}{the inverse of a sample covariance matrix}
}
\value{
a vector, which is the Global Minimum Variance Portfolio.
}
\description{
The functions computes the sample estimate of the weights of the global
minimum variance portfolio (see, e.g., Eq. (1.4) of
\insertCite{BODNAR21dynshrink;textual}{DOSPortfolio})).
}
