% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overlapping.R
\name{wGMVNonOverlapping}
\alias{wGMVNonOverlapping}
\title{The (non overlapping) Global Minimum Variance (GMV) portfolio using a dynamic optimal shrinkage scheme.}
\usage{
wGMVNonOverlapping(data, change_points, target_portfolio, relative_loss)
}
\arguments{
\item{data}{a matrix of size (n x p), where n>p, containing, for instance, log-returns.}

\item{change_points}{a vector of changepoints. The changepoints are what determines
when we recompute weights.}

\item{target_portfolio}{a vector which is the target weights that one wants to shrink to in the first period.}

\item{relative_loss}{a numeric of the initial value of the relative loss for the variance of the GMV portfolio.}
}
\value{
a matrix of shrunk GMV portfolio weights where each row corresponds to each change point.
}
\description{
This function implements the recursive estimation of the GMV portfolio in the dynamic optimal shrinkage setting.
The method demands that the data is on long format, observations are on rows and covariates are columns and that
the number of rows are greater than the number of columns. This function estimates each GMV portfolio
using the most recent data, e.g. from the last break point to the current.
}
\examples{
n <- 200*2
p <- 80
change_point <- c(199)
data <- matrix(rt(n*p, df=5), ncol=p, nrow=n)
target_portfolio <- as.vector(rep(1,p))/p
wGMVNonOverlapping(data, change_point, target_portfolio, 1)
}
\references{
\insertAllCited{}
}
\seealso{
section 2.1 \insertCite{BODNAR21dynshrink}{DOSPortfolio}
}
