% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{ComputeBeta}
\alias{ComputeBeta}
\title{A helper function for computing beta coefficients used in the case of the overlapping sample \insertCite{BODNAR21dynshrink}{DOSPortfolio}}
\usage{
ComputeBeta(i, j, Psi)
}
\arguments{
\item{i}{an integer greater than one.}

\item{j}{an integer greater than one.}

\item{Psi}{vector, the vector of the optimal shrinkage intensities computed in the previous step of the recursion.}
}
\value{
a number
}
\description{
The function computes the beta coefficients from Eq. (2.20) in \insertCite{BODNAR21dynshrink;textual}{DOSPortfolio},
which are used in the recursive computation of the dynamic shrinkage estimator of the GMV portfolio weights in the
case of overlapping samples.
}
\references{
\insertAllCited{}
}
\keyword{internal}
