% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOSPortfolio.R
\name{DOSPortfolio}
\alias{DOSPortfolio}
\title{The Dynamic Optimal Shrinkage Portfolio interface.}
\usage{
DOSPortfolio(
  data,
  change_points,
  relative_loss = NULL,
  target_w = OnesVec(ncol(data))/ncol(data),
  shrinkage_type = "non-overlapping"
)
}
\arguments{
\item{data}{a matrix of size (n x p), where n>p, containing, for instance, log-returns.}

\item{change_points}{a vector of break points. The breakpoints are what determines
when we recompute weights.}

\item{relative_loss}{a numeric of the initial value of the relative loss for the variance of the GMV portfolio.}

\item{target_w}{a vector which is the target weights that one wants to shrink to in the first period.}

\item{shrinkage_type}{the type of shrinkage estimator to use. The two implemented are "non-overlapping" and "overlapping"}
}
\value{
a DOSPortfolio class with the GMV weights and its corresponding type.
}
\description{
The Dynamic Optimal Shrinkage Portfolio interface.
}
\examples{
n <- 250*2
p <- 80
c <- p/n
change_points <- c(120, 240)
data <- matrix(rt(n*p, df=5), ncol=p, nrow=n)
weights <- DOSPortfolio(data, change_points, 1)

}
