% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r0Strategy}
\alias{r0Strategy}
\title{Computes the relative loss of the target portfolio used}
\usage{
r0Strategy(S_inv, S, target_portfolio, c)
}
\arguments{
\item{S_inv}{the inverse of a sample covariance matrix.}

\item{S}{a sample covariance matrix.}

\item{target_portfolio}{a vector which determines the weights of the target
portfolio used when the shrinkage estimator of the global minimum variance
portfolio is constructed for the first time.}

\item{c}{a numeric which is the concentration ratio.}
}
\value{
vector
}
\description{
The function computes the initial value of the relative loss in the variance
of the target portfolio as given in Eq. (2.10) of
\insertCite{BODNAR21dynshrink;textual}{DOSPortfolio}.
}
\references{
\insertAllCited{}
}
