% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{rUpdateOverlapping}
\alias{rUpdateOverlapping}
\title{The recursive estimation for updating the relative loss in the variance of the
holding portfolio.}
\usage{
rUpdateOverlapping(Psi, c, prev_R, K)
}
\arguments{
\item{Psi}{a vector of shrinkage coefficients.}

\item{c}{a numeric between 0 and 1. It is the concentration ration.}

\item{prev_R}{a numeric greater than 0, the previous value of the relative loss}

\item{K}{a numeric parameter.}
}
\value{
a number
}
\description{
This function implements equation (2.19), the recursive estimation for
updating the relative loss in the variance of the holding portfolio in the
overlapping scheme.
}
\keyword{internal}
