% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wGMV}
\alias{wGMV}
\title{The Global Minimum Variance (GMV) portfolio weights. This function
computes the global minimum portfolio weights which is part of Modern
Portfolio Theory, Markowitz (1952). For a positive semi-definite matrix
S, the weights are obtained as the solution to a quadratic optimization
problem given that the solution sums to one.}
\usage{
wGMV(S_inv)
}
\arguments{
\item{S_inv}{the inverse of a sample covariance matrix}
}
\value{
a vector, which is the Global Minimum Variance Portfolio.
}
\description{
The Global Minimum Variance (GMV) portfolio weights. This function
computes the global minimum portfolio weights which is part of Modern
Portfolio Theory, Markowitz (1952). For a positive semi-definite matrix
S, the weights are obtained as the solution to a quadratic optimization
problem given that the solution sums to one.
}
