% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapping.R
\name{wGMVOverlapping}
\alias{wGMVOverlapping}
\title{The (overlapping) Global Minimum Variance (GMV) portfolio using a dynamic optimal shrinkage scheme.}
\usage{
wGMVOverlapping(data, break_points, target_w, R)
}
\arguments{
\item{data}{a matrix of size (n x p), where n>p, containing, for instance, log-returns.}

\item{break_points}{a vector of break points. The breakpoints are what determines
when we recompute weights.}

\item{target_w}{a vector which is the target weights that one wants to shrink to in the first period.}

\item{R}{a numeric of the initial value of the relative loss for the variance of the GMV portfolio.}
}
\value{
vector of portfolio weights
}
\description{
This function implements the second version of the recursive estimation of the GMV portfolio in the dynamic optimal shrinkage setting.
The difference between this function and the function \code{wGMVNonOverlapping(...)} is that this function estimates each GMV portfolio using
all available data. To see a more detailed description of the method, see vigenette
\code{vignette("DOSPortfolio", package = "DOSPortfolio")}.
}
\examples{
n <- 200*2
p <- 80
c <- p/n
break_points <- c(199)
data <- matrix(rt(n*p, df=5), ncol=p, nrow=n)
target_w <- as.vector(rep(1,p))/p
wGMVOverlapping(data, break_points, target_w, 1)

}
